move orientation entropy to bearings module
